refactor(rag): extract shared ranking logic for vector search
